Fix build script and suppress chunk warning
