+++ README.md
@@ -4,8 +4,8 @@
 
 1. Install packages `npm i`
 2. Run dev server `npm run dev`
-4. Static assets are stored at `public/` and are avaliable at `/` on the web
-5. All TS is in `src/`
+3. Build for Cloudflare Workers `npm run build`
+4. Preview the build `npm run preview`
+5. Static assets are stored at `public/` and are avaliable at `/` on the web
 
 ### Create a new test
-
