--- index.html
@@ -1,21 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Does my browser support...</title>
-  </head>
-  <body>
-    <main>
-      <h1>Does your browser support...</h1>
-      <div class="browser-meta">
-        <div><strong>User agent:</strong> <span id="useragent"></span></div>
-        <div><strong>Detected browser:</strong> <span id="detectedBrowser"></span></div>
-        <div><strong>Latest stable version:</strong> <span id="latestBrowser"></span></div>
-      </div>
-      <div id="summary" class="summary"></div>
-      <ul class="tests"></ul>
-    </main>
-    <script type="module" src="/src/ts/main.ts"></script>
-  </body>
-</html>
