+++ package.json
@@ -4,28 +4,22 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "vite",
-    "build": "tsc && vite build",
-    "preview": "vite preview"
+    "dev": "vite dev",
+    "build": "npm exec -- svelte-kit sync && vite build",
+    "preview": "vite preview",
+    "check": "npm exec -- svelte-kit sync && svelte-check --tsconfig ./tsconfig.json"
   },
   "devDependencies": {
+    "@sveltejs/adapter-cloudflare": "^4.4.0",
+    "@sveltejs/kit": "^2.20.0",
+    "@sveltejs/vite-plugin-svelte": "^5.1.1",
     "@types/node": "^20.11.19",
-    "postcss": "^8.4.35",
-    "postcss-preset-env": "^9.3.0",
-    "sass": "^1.71.0",
+    "svelte": "^5.25.1",
+    "svelte-check": "^3.8.4",
     "typescript": "^5.2.2",
-    "vite": "^5.1.0"
+    "vite": "^6.0.0"
   },
   "dependencies": {
-    "@babel/core": "^7.23.9",
-    "@babel/preset-env": "^7.23.9",
-    "@babel/runtime": "^7.23.9",
-    "@rollup/plugin-babel": "^6.0.4",
-    "@shoelace-style/shoelace": "^2.14.0",
-    "@vitejs/plugin-legacy": "^5.3.0",
-    "babel": "^6.23.0",
-    "caniuse-lite": "^1.0.30001765",
-    "core-js": "^3.36.0",
-    "postcss-load-config": "^5.0.3"
+    "caniuse-lite": "^1.0.30001765"
   }
 }
