+++ src/app.css
@@ -1,17 +1,96 @@
-html, body, main {
+main h1 {
+    text-align: center;
+}
+
+main .browser-meta {
+    display: grid;
+    gap: 0.35em;
+    margin-bottom: 1em;
+}
+
+main .summary {
+    margin-bottom: 1.2em;
+    padding: 0.75em 1em;
+    border-radius: 0.5em;
+    background: #f1f3f5;
+    display: flex;
+    flex-wrap: wrap;
+    gap: 0.4em;
+    align-items: center;
+}
+
+main .separator {
+    opacity: 0.6;
+}
+
+main ul.tests {
+    list-style-type: none;
+    padding: 0;
+}
+
+main ul.tests li {
+    margin: 0.5em;
+    padding: 0.75em;
+    border-radius: 0.5em;
+    display: flex;
+    flex-direction: column;
+    gap: 0.4em;
+}
+
+main ul.tests li.passed {
+    background: #28a745;
+    color: #fff;
+}
+
+main ul.tests li.failed {
+    background: #dc3545;
+    color: #fff;
+}
+
+main ul.tests li.partial {
+    background: #f5c542;
+    color: #000;
+}
+
+main ul.tests li.unknown {
+    background: #6c757d;
+    color: #fff;
+}
+
+main ul.tests li .test-title {
+    font-weight: 600;
+}
+
+main ul.tests li .test-details {
+    font-size: 0.9em;
+}
+
+main ul.tests li .test-meta {
+    font-size: 0.85em;
+    display: flex;
+    flex-direction: column;
+    gap: 0.2em;
+}
+
+main ul.tests li .test-meta-item a {
+    color: inherit;
+    text-decoration: underline;
+}
+html,
+body,
+main {
     width: 100%;
     height: 100%;
     padding: 0;
     margin: 0;
-    // Issues with staying at the top
     position: absolute;
     top: 0;
     left: 0;
-    // color: oklch(61% 0.2 29);
-    font-family: 'IBM Plex Sans', system-ui,-apple-system,"Segoe UI",Roboto,"Helvetica Neue","Noto Sans","Liberation Sans",Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol","Noto Color Emoji";
+    font-family: 'IBM Plex Sans', system-ui, -apple-system, 'Segoe UI', Roboto, 'Helvetica Neue',
+        'Noto Sans', 'Liberation Sans', Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji',
+        'Segoe UI Symbol', 'Noto Color Emoji';
 }
 
-// Fonts
 @font-face {
     font-family: 'IBM Plex Sans';
     src: url('/fonts/woff2/IBMPlexSans-Regular.woff2') format('woff2'),
@@ -22,6 +101,7 @@ html, body, main {
     font-style: normal;
     font-display: swap;
 }
+
 @font-face {
     font-family: 'IBM Plex Sans';
     src: url('/fonts/woff2/IBMPlexSans-Bold.woff2') format('woff2'),
@@ -32,6 +112,7 @@ html, body, main {
     font-style: normal;
     font-display: swap;
 }
+
 @font-face {
     font-family: 'IBM Plex Sans';
     src: url('/fonts/woff2/IBMPlexSans-Italic.woff2') format('woff2'),
@@ -42,6 +123,7 @@ html, body, main {
     font-weight: normal;
     font-display: swap;
 }
+
 @font-face {
     font-family: 'IBM Plex Sans';
     src: url('/fonts/woff2/IBMPlexSans-BoldItalic.woff2') format('woff2'),
