+++ src/app.d.ts
@@ -1,0 +1,8 @@
+// See https://kit.svelte.dev/docs/types#app
+// for information about these interfaces
+
+declare global {
+    namespace App {}
+}
+
+export {};
