--- src/scss/index.scss
@@ -1,77 +1,0 @@
-@import 'global';
-
-main {
-    h1 {
-        text-align: center;
-    }
-
-    .browser-meta {
-        display: grid;
-        gap: 0.35em;
-        margin-bottom: 1em;
-    }
-
-    .summary {
-        margin-bottom: 1.2em;
-        padding: 0.75em 1em;
-        border-radius: 0.5em;
-        background: #f1f3f5;
-        display: flex;
-        flex-wrap: wrap;
-        gap: 0.4em;
-        align-items: center;
-    }
-
-    .separator {
-        opacity: 0.6;
-    }
-
-    ul.tests {
-        list-style-type: none;
-        padding: 0;
-        li {
-            margin: 0.5em;
-            padding: 0.75em;
-            border-radius: 0.5em;
-            display: flex;
-            flex-direction: column;
-            gap: 0.4em;
-            &.passed {
-                background: #28a745;
-                color: #fff;
-            }
-            &.failed {
-                background: #dc3545;
-                color: #fff;
-            }
-            &.partial {
-                background: #f5c542;
-                color: #000;
-            }
-            &.unknown {
-                background: #6c757d;
-                color: #fff;
-            }
-
-            .test-title {
-                font-weight: 600;
-            }
-
-            .test-details {
-                font-size: 0.9em;
-            }
-
-            .test-meta {
-                font-size: 0.85em;
-                display: flex;
-                flex-direction: column;
-                gap: 0.2em;
-            }
-
-            .test-meta-item a {
-                color: inherit;
-                text-decoration: underline;
-            }
-        }
-    }
-}
