--- src/ts/vite-env.d.ts
@@ -1,11 +1,0 @@
-/// <reference types="vite/client" />
-
-declare global {
-    interface Window {
-        webkitAudioContext?: typeof AudioContext;
-        webkitRTCPeerConnection?: typeof RTCPeerConnection;
-        mozRTCPeerConnection?: typeof RTCPeerConnection;
-    }
-}
-
-export {};
