+++ svelte.config.js
@@ -1,0 +1,11 @@
+import adapter from '@sveltejs/adapter-cloudflare';
+import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
+
+const config = {
+    preprocess: vitePreprocess(),
+    kit: {
+        adapter: adapter(),
+    },
+};
+
+export default config;
