+++ vite.config.js
@@ -1,10 +1,18 @@
-import legacy from '@vitejs/plugin-legacy';
-// import pluginPurgeCss from "@mojojoejo/vite-plugin-purgecss";
+import { sveltekit } from '@sveltejs/kit/vite';
+import { defineConfig } from 'vite';
 
-export default {
-  plugins: [
-    legacy({
-      targets: ['defaults', 'not IE 11'],
-    }),
-  ],
-}
+export default defineConfig({
+    plugins: [sveltekit()],
+    build: {
+        chunkSizeWarningLimit: 1500,
+        rollupOptions: {
+            output: {
+                manualChunks(id) {
+                    if (id.includes('node_modules/caniuse-lite')) {
+                        return 'caniuse';
+                    }
+                },
+            },
+        },
+    },
+});
